Big changes before rerun, read desc
Ok. this is trying to fix the fact that that last large run on the cluster didnt learn.
I think its because the next_x and next_y were swapped. Ive decided to make a bunch of changes including that one.

- Went back to the really old successor_agent.py, and changed action selection to not choose invalid moves

-changed mini_grid.py to do completely random generations without replacement

going to run again 5k episodes on the cluster to see if anything gets fixed. crossing fingers